Add files via upload
sensor package, can read a list of sensor readings from the computers hardware. 
I chose to read the CPU temp. Pulled via awk command 
also took the date and time of when the command is run

this can eventually be put into a graph of some kind for system maintenance